% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iconlist}
\alias{iconlist}
\title{Contains a list of valid icon names}
\format{An object of class \code{character} of length 596.}
\usage{
iconlist
}
\description{
Contains a list of valid icon names
}
\keyword{datasets}
