% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quadrants}
\alias{quadrants}
\title{The possible values for \code{towards} or \code{from}}
\format{An object of class \code{character} of length 16.}
\usage{
quadrants
}
\description{
The possible values for \code{towards} or \code{from}
}
\keyword{datasets}
