% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svglist}
\alias{svglist}
\title{Contains a list of valid svg's names}
\format{An object of class \code{character} of length 219.}
\usage{
svglist
}
\description{
Contains a list of valid svg's names
}
\keyword{datasets}
