% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{useWeatherIcons}
\alias{useWeatherIcons}
\title{useWeatherIcons}
\usage{
useWeatherIcons(wind = TRUE)
}
\arguments{
\item{wind}{Should the Wind-Stylesheet be included as dependency.
The default is \code{TRUE}}
}
\description{
Import the CSS files, if you're not using \code{\link{weathericon}}
  anywhere else in the UI.
}
