% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{weathericon}
\alias{weathericon}
\title{weathericon}
\usage{
weathericon(
  name = NULL,
  wind = TRUE,
  flip = NULL,
  rotate = NULL,
  towards = NULL,
  from = NULL,
  className = NULL,
  ...
)
}
\arguments{
\item{name}{Name of the icon}

\item{wind}{Should the Wind-Stylesheet be included as dependency.
The default is \code{TRUE}}

\item{flip}{Flip the icon. Either \code{horizontal} or \code{vertical}}

\item{rotate}{Rotate the icon. Valid values are \code{0,90,180,270,360}}

\item{towards}{The `towards`-direction for wind icons can be specified either by
degrees or by an orientation. See the details.}

\item{from}{The `from`-direction for wind icons can be specified either by
degrees or by an orientation. See the details.}

\item{className}{Define an additional classname for the icon tag.}

\item{...}{Additional arguments that are passed to the attributes
of the i-tag. Style arguments must be passed as a named list.}
}
\value{
A weather-icon element
}
\description{
Create a weathericon for use within a page. Icons can appear on
their own, inside of a button, or as an icon for a tabPanel() within a navbarPage().
}
\details{
The arguments \code{towards} and \code{from} can either be an integer value
between 0-360 or a character giving the direction. The possible items are
stored in the dataset \code{quadrants}.
}
\examples{
if (interactive()) {

weathericon("wi-forecast-io-fog")
weathericon("wi-wu-chanceflurries")
weathericon("wi-moon-waxing-crescent-3")
weathericon("wi-forecast-io-rain")
weathericon("storm-warning")
weathericon("day-fog")

## Rotating Icons
weathericon("wi wi-yahoo-7", rotate = "90")
weathericon("wi wi-yahoo-7", rotate = "180")
weathericon("wi wi-yahoo-7", rotate = "270")

## Flipping Icons
weathericon("wi wi-yahoo-7", flip = "vertical")

## Wind Icons with Directions
weathericon("wi wi-wind", towards = "sse")
weathericon("wi-wind", from = "sw")
weathericon("wi wi-wind", towards = 40)
weathericon("wi wi-wind", from = 40)

}
}
\seealso{
\href{https://erikflowers.github.io/weather-icons/}{Weather Icons}
}
